FP/requirements.txt
aiohttp==3.6.3
allure-pytest==2.9.42
allure-python-commons==2.9.42
apipkg==1.5
arlpy==1.7.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bokeh==2.2.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
decorator==4.4.2
discord==1.0.1
discord.py==1.5.1
execnet==1.8.1
Faker==8.5.1
greenlet==1.1.0
idna==2.8
imgkit==1.0.2
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
kiwisolver==1.3.1
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.4
packaging==20.4
pandas==1.1.4
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
pluggy==0.13.1
prompt-toolkit==3.0.10
py==1.10.0
Pygments==2.7.3
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
scipy==1.5.3
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.4.17
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.7
utm==0.6.0
wcwidth==0.2.5
webdriver-manager==3.4.2
xlrd==1.2.0
yarl==1.5.1
